cambio en archivos para que no accedan sin logueo

+++ admin/eliminar.php
@@ -1,14 +1,19 @@
 <?php
- include ("../controller/conexion.php");
- $id=$_GET['id'];
- $consulta = mysql_query("delete from celular where id_celular = '$id'");
- if($consulta){
- 	?>
- 	<div class="alert alert-success">El celular se eliminó correctamente</div>
- 	<?
- }else{
- 	?>
- 	<div class="alert alert-danger">El celular no se eliminó correctamente</div>
- 	<?
- }
+header("Content-Type: text/html;charset=utf-8");
+include('funciones.php');
+//uso de la funcion verificar_usuario()
+if (verificar_usuario()){
+	include ("../controller/conexion.php");
+ 	$id=$_GET['id'];
+ 	$consulta = mysql_query("delete from celular where id_celular = '$id'");
+ 	if($consulta){
+ 		?><div class="alert alert-success">El celular se eliminó correctamente</div><?
+ 	}else{
+ 		?><div class="alert alert-danger">El celular no se eliminó correctamente</div><?
+ 	}
+ } else {
+  //si el usuario no es verificado volvera al formulario de ingreso
+  ?><head><meta charset="utf-8"><?
+  echo "<script>alert(\"Por favor inicie su sesión.\"); location.href=\"../index.php\";</script>";
+}
 ?>
